+++ runtime/src/pallets/mod.rs
@@ -3,3 +3,6 @@ pub use system::*;
 
 pub mod timestamp;
 pub use timestamp::*;
+
+pub mod authorship;
+pub use authorship::*;
